Deliverable 1 : refactor and add parent pom
