+++ json-schema-store/src/main/java/uk/ac/ebi/biosamples/jsonschema/jsonschemastore/schema/document/MetaSchemaBlock.java
@@ -1,0 +1,12 @@
+package uk.ac.ebi.biosamples.jsonschema.jsonschemastore.schema.document;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+@Getter
+@Setter
+@Document
+public class MetaSchemaBlock {
+  private Object metaSchema;
+}
