+++ json-schema-store/src/main/java/uk/ac/ebi/biosamples/jsonschema/jsonschemastore/schema/repository/MetaSchemaRepository.java
@@ -1,0 +1,6 @@
+package uk.ac.ebi.biosamples.jsonschema.jsonschemastore.schema.repository;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+import uk.ac.ebi.biosamples.jsonschema.jsonschemastore.schema.document.MetaSchemaBlock;
+
+public interface MetaSchemaRepository extends MongoRepository<MetaSchemaBlock, String> {}
