+++ json-schema-store/src/main/java/uk/ac/ebi/biosamples/jsonschema/jsonschemastore/schema/service/MetaSchemaService.java
@@ -1,0 +1,4 @@
+package uk.ac.ebi.biosamples.jsonschema.jsonschemastore.schema.service;
+
+public class MetaSchemaService {
+}
