--- json-schema-store/src/main/java/uk/ac/ebi/biosamples/jsonschemastore/client/dto/ValidateResponseDocument.java
@@ -1,23 +1,0 @@
-package uk.ac.ebi.biosamples.jsonschemastore.client.dto;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-@EqualsAndHashCode
-public class ValidateResponseDocument {
-  private final List<String> errors;
-  private final String dataPath;
-
-  @JsonCreator(mode = JsonCreator.Mode.PROPERTIES)
-  public ValidateResponseDocument(
-      @JsonProperty("errors") final List<String> errors,
-      @JsonProperty("dataPath") final String dataPath) {
-    this.errors = errors;
-    this.dataPath = dataPath;
-  }
-}
