--- json-schema-store/src/main/java/uk/ac/ebi/biosamples/jsonschemastore/schema/document/MetaSchemaBlock.java
@@ -1,12 +1,0 @@
-package uk.ac.ebi.biosamples.jsonschemastore.schema.document;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-@Getter
-@Setter
-@Document
-public class MetaSchemaBlock {
-  private Object metaSchema;
-}
