--- json-schema-store/src/main/java/uk/ac/ebi/biosamples/jsonschemastore/schema/service/MetaSchemaService.java
@@ -1,4 +1,0 @@
-package uk.ac.ebi.biosamples.jsonschemastore.schema.service;
-
-public class MetaSchemaService {
-}
