--- json-schema-store/src/main/java/uk/ac/ebi/biosamples/jsonschemastore/schema/util/JsonSchemaMappingUtil.java
@@ -1,37 +1,0 @@
-package uk.ac.ebi.biosamples.jsonschemastore.schema.util;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class JsonSchemaMappingUtil {
-
-  public static <T> JsonNode convertSchemaBlockToJson(T schema) {
-    ObjectMapper objectMapper = new ObjectMapper();
-    ObjectNode jsonNode = objectMapper.valueToTree(schema);
-    jsonNode.remove("schemalessData");
-    return jsonNode;
-  }
-
-  public static <T> JsonNode convertObjectToJson(T object) {
-    ObjectMapper objectMapper = new ObjectMapper();
-    return objectMapper.valueToTree(object);
-  }
-
-  public static JsonNode getSchemaObject() {
-    ObjectMapper objectMapper = new ObjectMapper();
-    try {
-      return objectMapper.readTree(
-              "{\n    \"$schema\": \"http://json-schema.org/draft-07/schema#\",\n    \"$ref\": \"#/definitions/Welcome\",\n    \"definitions\": {\n        \"Welcome\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"$schema\": {\n                    \"type\": \"string\",\n                    \"format\": \"uri\",\n                    \"qt-uri-protocols\": [\n                        \"http\"\n                    ]\n                },\n                \"$id\": {\n                    \"type\": \"string\",\n                    \"format\": \"uri\",\n                    \"qt-uri-protocols\": [\n                        \"https\"\n                    ],\n                    \"qt-uri-extensions\": [\n                        \".0\"\n                    ]\n                },\n                \"title\": {\n                    \"type\": \"string\"\n                },\n                \"description\": {\n                    \"type\": \"string\"\n                },\n                \"type\": {\n                    \"type\": \"string\"\n                },\n                \"meta\": {\n                    \"$ref\": \"#/definitions/Meta\"\n                },\n                \"properties\": {\n                    \"$ref\": \"#/definitions/Properties\"\n                },\n                \"required\": {\n                    \"type\": \"array\",\n                    \"items\": {\n                        \"type\": \"string\"\n                    }\n                },\n                \"oneof\": {\n                    \"type\": \"array\",\n                    \"items\": {\n                        \"$ref\": \"#/definitions/Oneof\"\n                    }\n                },\n                \"additionalProperties\": {\n                    \"type\": \"boolean\"\n                },\n                \"examples\": {\n                    \"type\": \"array\",\n                    \"items\": {\n                        \"$ref\": \"#/definitions/WelcomeExample\"\n                    }\n                }\n            },\n            \"required\": [\n                \"$id\",\n                \"$schema\",\n                \"additionalProperties\",\n                \"description\",\n                \"examples\",\n                \"meta\",\n                \"oneof\",\n                \"properties\",\n                \"required\",\n                \"title\",\n                \"type\"\n            ],\n            \"title\": \"Welcome\"\n        },\n        \"WelcomeExample\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"term\": {\n                    \"$ref\": \"#/definitions/ClassOfOnset\"\n                },\n                \"classOfOnset\": {\n                    \"$ref\": \"#/definitions/ClassOfOnset\"\n                }\n            },\n            \"required\": [\n                \"classOfOnset\",\n                \"term\"\n            ],\n            \"title\": \"WelcomeExample\"\n        },\n        \"Meta\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"contributors\": {\n                    \"type\": \"array\",\n                    \"items\": {\n                        \"$ref\": \"#/definitions/Contributor\"\n                    }\n                },\n                \"provenance\": {\n                    \"type\": \"array\",\n                    \"items\": {\n                        \"$ref\": \"#/definitions/Contributor\"\n                    }\n                },\n                \"used_by\": {\n                    \"type\": \"array\",\n                    \"items\": {\n                        \"$ref\": \"#/definitions/Contributor\"\n                    }\n                },\n                \"sb_status\": {\n                    \"type\": \"string\"\n                }\n            },\n            \"required\": [\n                \"contributors\",\n                \"provenance\",\n                \"sb_status\",\n                \"used_by\"\n            ],\n            \"title\": \"Meta\"\n        },\n        \"Contributor\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"description\": {\n                    \"type\": \"string\"\n                },\n                \"id\": {\n                    \"type\": \"string\",\n                    \"qt-uri-protocols\": [\n                        \"https\"\n                    ],\n                    \"qt-uri-extensions\": [\n                        \".rst\"\n                    ]\n                }\n            },\n            \"required\": [\n                \"description\"\n            ],\n            \"title\": \"Contributor\"\n        },\n        \"Oneof\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"properties\": {\n                    \"type\": \"array\",\n                    \"items\": {\n                        \"type\": \"string\"\n                    }\n                }\n            },\n            \"required\": [\n                \"properties\"\n            ],\n            \"title\": \"Oneof\"\n        },\n        \"Properties\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"term\": {\n                    \"$ref\": \"#/definitions/Term\"\n                },\n                \"ageOfOnset\": {\n                    \"$ref\": \"#/definitions/AgeOfOnset\"\n                },\n                \"ageRangeOfOnset\": {\n                    \"$ref\": \"#/definitions/AgeRangeOfOnset\"\n                },\n                \"classOfOnset\": {\n                    \"$ref\": \"#/definitions/PropertiesClassOfOnset\"\n                },\n                \"diseaseStage\": {\n                    \"$ref\": \"#/definitions/DiseaseStage\"\n                },\n                \"tnmFinding\": {\n                    \"$ref\": \"#/definitions/DiseaseStage\"\n                }\n            },\n            \"required\": [\n                \"ageOfOnset\",\n                \"ageRangeOfOnset\",\n                \"classOfOnset\",\n                \"diseaseStage\",\n                \"term\",\n                \"tnmFinding\"\n            ],\n            \"title\": \"Properties\"\n        },\n        \"AgeOfOnset\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"allof\": {\n                    \"type\": \"array\",\n                    \"items\": {\n                        \"$ref\": \"#/definitions/AgeOfOnsetAllof\"\n                    }\n                }\n            },\n            \"required\": [\n                \"allof\"\n            ],\n            \"title\": \"AgeOfOnset\"\n        },\n        \"AgeOfOnsetAllof\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"$ref\": {\n                    \"type\": \"string\",\n                    \"format\": \"uri\",\n                    \"qt-uri-protocols\": [\n                        \"https\"\n                    ],\n                    \"qt-uri-extensions\": [\n                        \".json\"\n                    ]\n                },\n                \"description\": {\n                    \"type\": \"string\"\n                },\n                \"examples\": {\n                    \"type\": \"array\",\n                    \"items\": {\n                        \"$ref\": \"#/definitions/Start\"\n                    }\n                }\n            },\n            \"required\": [],\n            \"title\": \"AgeOfOnsetAllof\"\n        },\n        \"Start\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"age\": {\n                    \"type\": \"string\"\n                }\n            },\n            \"required\": [\n                \"age\"\n            ],\n            \"title\": \"Start\"\n        },\n        \"AgeRangeOfOnset\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"description\": {\n                    \"type\": \"string\"\n                },\n                \"$ref\": {\n                    \"type\": \"string\",\n                    \"format\": \"uri\",\n                    \"qt-uri-protocols\": [\n                        \"https\"\n                    ],\n                    \"qt-uri-extensions\": [\n                        \".json\"\n                    ]\n                },\n                \"examples\": {\n                    \"type\": \"array\",\n                    \"items\": {\n                        \"$ref\": \"#/definitions/AgeRangeOfOnsetExample\"\n                    }\n                }\n            },\n            \"required\": [\n                \"$ref\",\n                \"description\",\n                \"examples\"\n            ],\n            \"title\": \"AgeRangeOfOnset\"\n        },\n        \"AgeRangeOfOnsetExample\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"start\": {\n                    \"$ref\": \"#/definitions/Start\"\n                }\n            },\n            \"required\": [\n                \"start\"\n            ],\n            \"title\": \"AgeRangeOfOnsetExample\"\n        },\n        \"PropertiesClassOfOnset\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"description\": {\n                    \"type\": \"string\"\n                },\n                \"$ref\": {\n                    \"type\": \"string\",\n                    \"format\": \"uri\",\n                    \"qt-uri-protocols\": [\n                        \"https\"\n                    ],\n                    \"qt-uri-extensions\": [\n                        \".json\"\n                    ]\n                },\n                \"examples\": {\n                    \"type\": \"array\",\n                    \"items\": {\n                        \"$ref\": \"#/definitions/ClassOfOnset\"\n                    }\n                }\n            },\n            \"required\": [\n                \"$ref\",\n                \"description\",\n                \"examples\"\n            ],\n            \"title\": \"PropertiesClassOfOnset\"\n        },\n        \"DiseaseStage\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"description\": {\n                    \"type\": \"string\"\n                },\n                \"type\": {\n                    \"type\": \"string\"\n                },\n                \"items\": {\n                    \"$ref\": \"#/definitions/Items\"\n                },\n                \"examples\": {\n                    \"type\": \"array\",\n                    \"items\": {\n                        \"type\": \"array\",\n                        \"items\": {\n                            \"$ref\": \"#/definitions/ClassOfOnset\"\n                        }\n                    }\n                }\n            },\n            \"required\": [\n                \"description\",\n                \"examples\",\n                \"items\",\n                \"type\"\n            ],\n            \"title\": \"DiseaseStage\"\n        },\n        \"ClassOfOnset\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"id\": {\n                    \"type\": \"string\"\n                },\n                \"label\": {\n                    \"type\": \"string\"\n                }\n            },\n            \"required\": [\n                \"id\",\n                \"label\"\n            ],\n            \"title\": \"ClassOfOnset\"\n        },\n        \"Items\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"$ref\": {\n                    \"type\": \"string\",\n                    \"format\": \"uri\",\n                    \"qt-uri-protocols\": [\n                        \"https\"\n                    ],\n                    \"qt-uri-extensions\": [\n                        \".json\"\n                    ]\n                }\n            },\n            \"required\": [\n                \"$ref\"\n            ],\n            \"title\": \"Items\"\n        },\n        \"Term\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"allof\": {\n                    \"type\": \"array\",\n                    \"items\": {\n                        \"$ref\": \"#/definitions/TermAllof\"\n                    }\n                }\n            },\n            \"required\": [\n                \"allof\"\n            ],\n            \"title\": \"Term\"\n        },\n        \"TermAllof\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"$ref\": {\n                    \"type\": \"string\",\n                    \"format\": \"uri\",\n                    \"qt-uri-protocols\": [\n                        \"https\"\n                    ],\n                    \"qt-uri-extensions\": [\n                        \".json\"\n                    ]\n                },\n                \"description\": {\n                    \"type\": \"string\"\n                },\n                \"examples\": {\n                    \"type\": \"array\",\n                    \"items\": {\n                        \"$ref\": \"#/definitions/AllofExample\"\n                    }\n                }\n            },\n            \"required\": [],\n            \"title\": \"TermAllof\"\n        },\n        \"AllofExample\": {\n            \"type\": \"object\",\n            \"additionalProperties\": false,\n            \"properties\": {\n                \"id\": {\n                    \"type\": \"string\"\n                }\n            },\n            \"required\": [\n                \"id\"\n            ],\n            \"title\": \"AllofExample\"\n        }\n    }\n}");
-    } catch (JsonProcessingException e) {
-      log.error("Error! ", e);
-      return null;
-    }
-  }
-}
