+++ json-schema-store/src/test/java/uk/ac/ebi/biosamples/jsonschema/jsonschemastore/JsonSchemaStoreApplicationTests.java
@@ -1,0 +1,13 @@
+package uk.ac.ebi.biosamples.jsonschema.jsonschemastore;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class JsonSchemaStoreApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
