--- json-schema-store/src/test/java/uk/ac/ebi/biosamples/jsonschemastore/JsonSchemaStoreApplicationTests.java
@@ -1,13 +1,0 @@
-package uk.ac.ebi.biosamples.jsonschemastore;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class JsonSchemaStoreApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
